Move wagmi connector in a separate components to be user by storybook components
