+++ src/components/web3/Web3Provider.tsx
@@ -1,0 +1,54 @@
+import { WagmiConfig, createClient, configureChains, goerli, mainnet } from 'wagmi'
+import { hardhat } from 'wagmi/chains'
+
+// import { alchemyProvider } from 'wagmi/providers/alchemy'
+import { publicProvider } from 'wagmi/providers/public'
+
+import { CoinbaseWalletConnector } from 'wagmi/connectors/coinbaseWallet'
+import { InjectedConnector } from 'wagmi/connectors/injected'
+import { MetaMaskConnector } from 'wagmi/connectors/metaMask'
+import { WalletConnectConnector } from 'wagmi/connectors/walletConnect'
+
+interface Web3ProviderProps {
+  children: React.ReactNode
+}
+
+const Web3Provider: React.FC<Web3ProviderProps> = ({ children }) => {
+  const { chains, provider, webSocketProvider } = configureChains(
+    [mainnet, hardhat, goerli],
+    [publicProvider()], // alchemyProvider({ apiKey: 'yourAlchemyApiKey' }),
+  )
+
+  // Set up client
+  const client = createClient({
+    autoConnect: true,
+    connectors: [
+      new MetaMaskConnector({ chains }),
+      new CoinbaseWalletConnector({
+        chains,
+        options: {
+          appName: 'wagmi',
+        },
+      }),
+      new WalletConnectConnector({
+        chains,
+        options: {
+          qrcode: true,
+        },
+      }),
+      new InjectedConnector({
+        chains,
+        options: {
+          name: 'Injected',
+          shimDisconnect: true,
+        },
+      }),
+    ],
+    provider,
+    webSocketProvider,
+  })
+
+  return <WagmiConfig client={client}>{children}</WagmiConfig>
+}
+
+export default Web3Provider
