+++ src/pages/_app.tsx
@@ -2,67 +2,22 @@ import React from 'react'
 import { AppProps } from 'next/app'
 import { ChakraProvider } from '@chakra-ui/react'
 import { ColorModeScript } from '@chakra-ui/color-mode'
-import { WagmiConfig, createClient, configureChains, goerli, mainnet } from 'wagmi'
-import { hardhat } from 'wagmi/chains'
-
-// import { alchemyProvider } from 'wagmi/providers/alchemy'
-import { publicProvider } from 'wagmi/providers/public'
-
-import { CoinbaseWalletConnector } from 'wagmi/connectors/coinbaseWallet'
-import { InjectedConnector } from 'wagmi/connectors/injected'
-import { MetaMaskConnector } from 'wagmi/connectors/metaMask'
-import { WalletConnectConnector } from 'wagmi/connectors/walletConnect'
 
 import Header from '../config'
+import Web3Provider from '../components/web3/Web3Provider'
 import theme from '../styles/theme'
 import Layout from '../components/dom/Layout'
-
-// Configure chains & providers with the Alchemy provider.
-// Two popular providers are Alchemy (alchemy.com) and Infura (infura.io)
-const { chains, provider, webSocketProvider } = configureChains(
-  [mainnet, hardhat, goerli],
-  [publicProvider()], // alchemyProvider({ apiKey: 'yourAlchemyApiKey' }),
-)
-
-// Set up client
-const client = createClient({
-  autoConnect: true,
-  connectors: [
-    new MetaMaskConnector({ chains }),
-    new CoinbaseWalletConnector({
-      chains,
-      options: {
-        appName: 'wagmi',
-      },
-    }),
-    new WalletConnectConnector({
-      chains,
-      options: {
-        qrcode: true,
-      },
-    }),
-    new InjectedConnector({
-      chains,
-      options: {
-        name: 'Injected',
-        shimDisconnect: true,
-      },
-    }),
-  ],
-  provider,
-  webSocketProvider,
-})
 
 const App: React.FC<AppProps> = ({ Component, pageProps = { title: 'index' } }) => {
   return (
     <ChakraProvider theme={theme}>
       <ColorModeScript initialColorMode='light' />
       <Header title={pageProps.title} />
-      <WagmiConfig client={client}>
+      <Web3Provider>
         <Layout>
           <Component {...pageProps} />
         </Layout>
-      </WagmiConfig>
+      </Web3Provider>
     </ChakraProvider>
   )
 }
